scope/requirements.txt
flask==2.3.2
flask-cors==4.0.0
langchain==0.1.16
langchain-community==0.0.34
langchain-huggingface==0.0.3
langchain-text-splitters==0.0.1
sentence-transformers==2.7.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
PyPDF2==3.0.1
pdfplumber==0.10.0
requests==2.31.0
pyyaml==6.0
transformers>=4.39.0
torch==2.0.1
faiss-cpu==1.7.4
chromadb==0.4.6